# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CCMS API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/bookings/{booking_id}/pay": {
      "post": {
        "tags": [
          "bookings"
        ],
        "summary": "Pay Booking",
        "operationId": "pay_booking_api_v1_bookings__booking_id__pay_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Booking Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingPay"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/shop/orders/{order_id}/pay": {
      "post": {
        "tags": [
          "shop"
        ],
        "summary": "Pay Order",
        "operationId": "pay_order_api_v1_shop_orders__order_id__pay_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderPay"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopOrderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/bar/orders/{order_id}/pay": {
      "post": {
        "tags": [
          "bar"
        ],
        "summary": "Pay Order",
        "operationId": "pay_order_api_v1_bar_orders__order_id__pay_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BarPay"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BarOrderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/payments": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "List Payments",
        "operationId": "list_payments_api_v1_payments_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "source_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SourceType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Type"
            }
          },
          {
            "name": "method",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PaymentMethod"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Method"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/mine": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "My Payments",
        "operationId": "my_payments_api_v1_payments_mine_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/{payment_id}/refund": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Refund",
        "operationId": "refund_api_v1_payments__payment_id__refund_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Payment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefundRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/payments.csv": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "Payments Csv",
        "operationId": "payments_csv_api_v1_reports_payments_csv_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/tax-summary": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "Tax Summary",
        "operationId": "tax_summary_api_v1_reports_tax_summary_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaxSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit-logs": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "List Audit Logs",
        "description": "SRS 3.1 gives the audit trail to the OWNER alone, read-only. It is never written here.",
        "operationId": "list_audit_logs_api_v1_audit_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "name": "entity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity"
            }
          },
          {
            "name": "actor_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actor Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BarOrderOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "table_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Table Id"
          },
          "member_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Id"
          },
          "guest_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guest Name"
          },
          "kitchen_status": {
            "$ref": "#/components/schemas/KitchenStatus"
          },
          "payment_status": {
            "$ref": "#/components/schemas/BarPaymentStatus"
          },
          "is_tab": {
            "type": "boolean",
            "title": "Is Tab"
          },
          "subtotal_paise": {
            "type": "integer",
            "title": "Subtotal Paise"
          },
          "discount_paise": {
            "type": "integer",
            "title": "Discount Paise"
          },
          "total_paise": {
            "type": "integer",
            "title": "Total Paise"
          },
          "tax_paise": {
            "type": "integer",
            "title": "Tax Paise"
          },
          "paid_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paid At"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BarLineOut"
            },
            "type": "array",
            "title": "Items",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "table_id",
          "member_id",
          "guest_name",
          "kitchen_status",
          "payment_status",
          "is_tab",
          "subtotal_paise",
          "discount_paise",
          "total_paise",
          "tax_paise",
          "paid_at"
        ],
        "title": "BarOrderOut"
      },
      "BarPay": {
        "properties": {
          "method": {
            "$ref": "#/components/schemas/PaymentMethod"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "method"
        ],
        "title": "BarPay"
      },
...
      "BookingOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "court_id": {
            "type": "integer",
            "title": "Court Id"
          },
          "member_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Id"
          },
          "guest_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guest Name"
          },
          "start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Start At"
          },
          "end_at": {
            "type": "string",
            "format": "date-time",
            "title": "End At"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "tier_applied": {
            "$ref": "#/components/schemas/Tier"
          },
          "price_paise": {
            "type": "integer",
            "title": "Price Paise"
          },
          "payment_status": {
            "$ref": "#/components/schemas/BookingPaymentStatus"
          },
          "source": {
            "$ref": "#/components/schemas/BookingSource"
          },
          "member_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Name"
          },
          "member_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Code"
          }
        },
        "type": "object",
        "required": [
          "id",
          "court_id",
          "member_id",
          "guest_name",
          "start_at",
          "end_at",
          "status",
          "tier_applied",
          "price_paise",
          "payment_status",
          "source"
        ],
        "title": "BookingOut"
      },
      "BookingPay": {
        "properties": {
          "payment_method": {
            "$ref": "#/components/schemas/PaymentMethod"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "payment_method"
        ],
        "title": "BookingPay"
      },
      "BookingPaymentStatus": {
        "type": "string",
        "enum": [
          "PAID",
          "UNPAID",
          "WAIVED",
          "REFUNDED"
        ],
        "title": "BookingPaymentStatus"
      },
      "BookingSource": {
        "type": "string",
        "enum": [
          "WEB",
          "FRONT_DESK",
          "PHONE"
        ],
        "title": "BookingSource"
      },
...
      "BookingStatus": {
        "type": "string",
        "enum": [
          "CONFIRMED",
          "CANCELLED",
          "COMPLETED",
          "NO_SHOW"
        ],
        "title": "BookingStatus"
      },
...
      "Fulfilment": {
        "type": "string",
        "enum": [
          "INSTORE",
          "PICKUP",
          "DELIVERY"
        ],
        "title": "Fulfilment"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderLineOut": {
        "properties": {
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "qty": {
            "type": "integer",
            "title": "Qty"
          },
          "unit_price_paise": {
            "type": "integer",
            "title": "Unit Price Paise"
          },
          "line_total_paise": {
            "type": "integer",
            "title": "Line Total Paise"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "name",
          "qty",
          "unit_price_paise",
          "line_total_paise"
        ],
        "title": "OrderLineOut"
      },
      "OrderPay": {
        "properties": {
          "payment_method": {
            "$ref": "#/components/schemas/PaymentMethod"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "payment_method"
        ],
        "title": "OrderPay"
      },
      "OrderPaymentStatus": {
        "type": "string",
        "enum": [
          "PAID",
          "UNPAID",
          "REFUNDED"
        ],
        "title": "OrderPaymentStatus"
      },
      "PageOut": {
        "properties": {
          "items": {
            "items": {},
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "PageOut"
      },
...
      "PaymentMethod": {
        "type": "string",
        "enum": [
          "CASH",
          "CARD",
          "UPI",
          "ONLINE_MOCK"
        ],
        "title": "PaymentMethod"
      },
      "PaymentOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType"
          },
          "source_id": {
            "type": "integer",
            "title": "Source Id"
          },
          "member_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Id"
          },
          "amount_paise": {
            "type": "integer",
            "title": "Amount Paise"
          },
          "tax_paise": {
            "type": "integer",
            "title": "Tax Paise"
          },
          "method": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "status": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          },
          "received_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_type",
          "source_id",
          "member_id",
          "amount_paise",
          "tax_paise",
          "method",
          "status",
          "reference",
          "received_by",
          "created_at"
        ],
        "title": "PaymentOut"
      },
...
      "RefundRequest": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RefundRequest"
      },
...
      "ShopChannel": {
        "type": "string",
        "enum": [
          "COUNTER",
          "ONLINE"
        ],
        "title": "ShopChannel"
      },
...
      "ShopOrderOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "member_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Id"
          },
          "guest_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guest Name"
          },
          "channel": {
            "$ref": "#/components/schemas/ShopChannel"
          },
          "fulfilment": {
            "$ref": "#/components/schemas/Fulfilment"
          },
          "delivery_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Address"
          },
          "status": {
            "$ref": "#/components/schemas/ShopOrderStatus"
          },
          "subtotal_paise": {
            "type": "integer",
            "title": "Subtotal Paise"
          },
          "discount_paise": {
            "type": "integer",
            "title": "Discount Paise"
          },
          "total_paise": {
            "type": "integer",
            "title": "Total Paise"
          },
          "tax_paise": {
            "type": "integer",
            "title": "Tax Paise"
          },
          "payment_status": {
            "$ref": "#/components/schemas/OrderPaymentStatus"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderLineOut"
            },
            "type": "array",
            "title": "Items",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "member_id",
          "guest_name",
          "channel",
          "fulfilment",
          "delivery_address",
          "status",
          "subtotal_paise",
          "discount_paise",
          "total_paise",
          "tax_paise",
          "payment_status"
        ],
        "title": "ShopOrderOut"
      },
      "ShopOrderStatus": {
        "type": "string",
        "enum": [
          "PLACED",
          "READY",
          "OUT_FOR_DELIVERY",
          "COMPLETED",
          "CANCELLED"
        ],
        "title": "ShopOrderStatus"
      },
...
      "SourceType": {
        "type": "string",
        "enum": [
          "BOOKING",
          "SOCIAL",
          "SHOP_ORDER",
          "BAR_ORDER",
          "MEMBERSHIP",
          "INVOICE"
        ],
        "title": "SourceType"
      },
...
      "TaxSummary": {
        "properties": {
          "month": {
            "type": "string",
            "title": "Month"
          },
          "revenue_paise": {
            "type": "integer",
            "title": "Revenue Paise"
          },
          "tax_paise": {
            "type": "integer",
            "title": "Tax Paise"
          },
          "by_source": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "integer"
              },
              "type": "object"
            },
            "type": "object",
            "title": "By Source"
          }
        },
        "type": "object",
        "required": [
          "month",
          "revenue_paise",
          "tax_paise",
          "by_source"
        ],
        "title": "TaxSummary"
      },
      "Tier": {
        "type": "string",
        "enum": [
          "GOLD",
          "SILVER",
          "JUNIOR",
          "WALKIN"
        ],
        "title": "Tier"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}